fix: overflow and sticky heading
